+++ src/ell/evaluation/evaluation.py
@@ -118,7 +118,7 @@ class Evaluation(LabelListMixin):
             self.labels.append(
                 Labeler(name="criterion", type=EvaluationLabelerType.CRITERION, label=criterion)
             )
-        assert len(self.labels) > 0, "No labels found, labeless evaluations coming soon!"
+        assert len(self.labels) > 0, "No labels found, labeless evaluations coming soon! Specify metrics, annotations, or criterion."
         assert not annotations, "Annotations are not supported yet."
